Update class structure
